Add travis-ci yml file for swift builds
This .travis.yml has 3 stages, The first stage is for builds scheduled
by cron, this job uploads a nightly artifact to an s3 bucket. The next
stage triggers when the repo is tagged and this uploads an artifact
named after the tag to s3, and the third stage of the build triggers on
branch updates and pull requests, and does not archive any artifacts.

Signed-off-by: Richard Berg <rberg@bitwise.io>